+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 
 * Added the user profiles feature interface, and it is accessible through "Countly::instance()::userProfile()" call.
 * Crash Reporting interface added and accessible through "Countly::instance()::crash()" call.
-* 
+* Added "disableUnhandledCrashReporting" function to the "Config" class to disable automatic uncaught crash reporting.
+* Added "setTotalBreadcrumbsAllowed(int)" function to the "Config" class to change allowed max breadcrumb count.
+
 * Fixed a bug where setting custom user properties would not work.
 * Fixed a bug where setting organization of the user would not work.